Assignment-2
Write a program that implements a slow [𝑶(𝒏𝟐)] sort method and a fast
[𝑶(𝒏𝒍𝒐𝒈𝒏)] sort method. Experiment with your program by running
your sort methods using 𝒏 random numbers as inputs for 𝒏 = 𝟏𝟎𝟎𝟎,𝒏
= 𝟏𝟎𝟎𝟎𝟎,𝒏 = 𝟏𝟎𝟎𝟎𝟎𝟎. Compare the running times by
implementing the timer function of the language you are using. (Bonus)
Use Binary Search on sorted data for each 𝒏 and report run times. (You
may search a random value in the range you are using)